main/requirements.txt
audioread==3.0.1
basic-pitch==0.3.3
Brotli==1.1.0
certifi==2024.2.2
cffi==1.16.0
charset-normalizer==3.3.2
colorama==0.4.6
coloredlogs==15.0.1
decorator==4.4.2
Deprecated==1.2.14
epub2pdf==0.1.0
flatbuffers==24.3.25
future==1.0.0
humanfriendly==10.0
idna==3.7
imageio==2.34.2
imageio-ffmpeg==0.5.1
img2pdf==0.5.1
joblib==1.4.2
lazy_loader==0.4
librosa==0.10.2.post1
llvmlite==0.42.0
loguru==0.6.0
lxml==5.2.2
mido==1.3.2
mir_eval==0.7
mobi==0.3.3
moviepy==1.0.3
mpmath==1.3.0
msgpack==1.0.8
mutagen==1.47.0
numba==0.59.1
numpy==1.26.4
onnxruntime==1.18.0
packaging==23.2
pikepdf==9.0.0
pillow==10.3.0
platformdirs==4.2.2
pooch==1.8.1
pretty_midi==0.2.10
proglog==0.1.10
protobuf==5.27.0
pycparser==2.22
pycryptodomex==3.20.0
pygame==2.5.2
pynput==1.7.7
pyo==1.0.5
PyOpenGL==3.1.7
PyOpenGL-accelerate==3.1.7
pyreadline3==3.4.1
python-rtmidi==1.5.8
requests==2.32.2
resampy==0.4.2
scikit-learn==1.5.0
scipy==1.13.1
six==1.16.0
soundfile==0.12.1
soxr==0.3.7
sympy==1.12
threadpoolctl==3.5.0
tqdm==4.66.4
typing_extensions==4.12.0
urllib3==2.2.1
websockets==12.0
win32-setctime==1.1.0
wrapt==1.16.0
wxPython==4.2.1
yt-dlp==2024.4.9
